Initial potential editor drafts of core and sparql
These will contain various respec errors esp due to unreferenced definitions. This is intentional as these documents are the minimal starting point from the original spec (which was a combination of the two separate documents and thus has all definitions in a single place). My goal here is to make sure that the initial documents do not have any content changes from the 1.0 starting point, as any such changes need to go through the proper W3C process.