#429: Have SHACL-SHACL warn about unsatisfiable usage of sh:in and sh:xone
This patch revises part of the implementation in PR 417.

* `in-minListLength` was not added to the Core document, but it was
  referenced in SHACL-SHACL.  Verbiage is added to Core in line with the
  text drafted for Issue 429.
* `xone-memberShape` was added as a reference in SHACL-SHACL, but
  doesn't exist in the Core document.  `xone-members-node` is the anchor
  in the Core document.  `xone-members-node` is now used.
* The `PropertyShape` for `sh:in` set a minimum required length of 1,
  which after discussion in Issue 429 appears to be a backward-
  incompatible change.  A shape paralleling the original suggestion for
  `sh:xone` is added to house that new restriction at the same warning-
  level severity.

An implementation style point: The property shapes housing
`sh:minListLength` are given IRIs after sketching a test
(`.../in-003.ttl`) and assuming it would be more desirable to have an
IRI-identified shape in the results graph, rather than a blank node.

A testing style point: Rather than running SHACL-SHACL over the whole of
a test, shapes from SHACL-SHACL are copied into some focused tests
(`.../{in,xone}-003.ttl`).

References:
* https://github.com/w3c/data-shapes/issues/429
* https://github.com/w3c/data-shapes/pull/417

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>